requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
fastapi==0.115.14
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
openai==1.13.3
python-dotenv==1.0.1
python-docx==1.1.0
reportlab==4.1.0
docx2pdf==0.1.8
faiss-cpu==1.7.4
sentence-transformers==2.7.0
